+++ src/Stryker.Core/Stryker.Core.UnitTest/DiffProviders/GitDiffProviderTests.cs
@@ -39,7 +39,7 @@ namespace Stryker.Core.UnitTest.DiffProviders
         public void ScanDiffReturnsListofFiles()
         {
             // Arrange
-            var basePath = FilePathUtils.NormalizePathSeparators("C://Users/JohnDoe/Project/Tests");
+            var basePath = FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests");
             var options = new StrykerOptions(gitDiffTarget: "d670460b4b4aece5915caf5c68d12f560a9fe3e4", basePath: basePath, fileSystem: new MockFileSystem());
 
             var gitInfoMock = new Mock<IGitInfoProvider>();
@@ -85,7 +85,7 @@ namespace Stryker.Core.UnitTest.DiffProviders
             gitInfoMock.Setup(x => x.DetermineCommit()).Returns(commitMock.Object);
 
             gitInfoMock.SetupGet(x => x.Repository).Returns(repositoryMock.Object);
-            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns("C:/Path/To/Repo");
+            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns("/c/Path/To/Repo");
             var target = new GitDiffProvider(options, gitInfoMock.Object);
 
             // Act
@@ -135,6 +135,290 @@ namespace Stryker.Core.UnitTest.DiffProviders
             var target = new GitDiffProvider(options, gitInfoMock.Object);
 
             Should.Throw<StrykerInputException>(() => target.ScanDiff());
+        }
+
+        [Fact]
+        public void ScanDiffReturnsListofFiles_ExcludingTestFilesInDiffIgnoreFiles()
+        {
+            // Arrange
+            var diffIgnoreFiles = new[] { "/c/Users/JohnDoe/Project/Tests/Test.cs" };
+
+            var basePath = FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests");
+            var options = new StrykerOptions(gitDiffTarget: "d670460b4b4aece5915caf5c68d12f560a9fe3e4", basePath: basePath, fileSystem: new MockFileSystem(), diffIgnoreFiles: diffIgnoreFiles);
+
+            var gitInfoMock = new Mock<IGitInfoProvider>();
+            var repositoryMock = new Mock<IRepository>(MockBehavior.Loose);
+            var commitMock = new Mock<Commit>(MockBehavior.Loose);
+            var branchMock = new Mock<Branch>(MockBehavior.Strict);
+            var patchMock = new Mock<Patch>(MockBehavior.Strict);
+            var patchEntryChangesMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+            var patchEntryChangesGitIgnoreMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+
+            // Setup of mocks
+            commitMock
+                .SetupGet(x => x.Tree)
+                .Returns(new Mock<Tree>(MockBehavior.Loose).Object);
+
+            branchMock
+                .SetupGet(x => x.Tip)
+                .Returns(commitMock.Object);
+
+            branchMock.SetupGet(x => x.UpstreamBranchCanonicalName).Returns("origin/branch");
+            branchMock.SetupGet(x => x.CanonicalName).Returns("refs/heads/branch");
+            branchMock.SetupGet(x => x.FriendlyName).Returns("branch");
+
+            repositoryMock
+                .Setup(x => x.Branches.GetEnumerator())
+                .Returns(new List<Branch> { branchMock.Object }.GetEnumerator())
+                .Verifiable();
+
+            patchEntryChangesMock
+                .SetupGet(x => x.Path)
+                .Returns("file.cs");
+
+            patchEntryChangesGitIgnoreMock
+                .SetupGet(x => x.Path)
+                .Returns(FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests/Test.cs"));
+
+            patchMock
+                .Setup(x => x.GetEnumerator())
+                .Returns(((IEnumerable<PatchEntryChanges>)new List<PatchEntryChanges>
+                {
+                    patchEntryChangesMock.Object,
+                    patchEntryChangesGitIgnoreMock.Object
+                }).GetEnumerator());
+
+            repositoryMock
+                .Setup(x => x.Diff.Compare<Patch>(It.IsAny<Tree>(), DiffTargets.WorkingDirectory))
+                .Returns(patchMock.Object);
+
+            repositoryMock
+                .Setup(x => x.Lookup(It.IsAny<ObjectId>())).Returns(commitMock.Object);
+
+            gitInfoMock.Setup(x => x.DetermineCommit()).Returns(commitMock.Object);
+
+            gitInfoMock.SetupGet(x => x.Repository).Returns(repositoryMock.Object);
+            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns("/c/Path/To/Repo");
+            var target = new GitDiffProvider(options, gitInfoMock.Object);
+
+            // Act
+            var res = target.ScanDiff();
+
+            // Assert
+            res.ChangedTestFiles.Count().ShouldBe(0);
+            res.ChangedSourceFiles.Count().ShouldBe(1);
+        }
+
+        [Fact]
+        public void ScanDiffReturnsListofFiles_ExcludingTestFilesInDiffIgnoreFiles_Single_Asterisk()
+        {
+            // Arrange
+            var diffIgnoreFiles = new[] { "/c/Users/JohnDoe/Project/*/Test.cs" };
+
+            var basePath = FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests");
+            var options = new StrykerOptions(gitDiffTarget: "d670460b4b4aece5915caf5c68d12f560a9fe3e4", basePath: basePath, fileSystem: new MockFileSystem(), diffIgnoreFiles: diffIgnoreFiles);
+
+            var gitInfoMock = new Mock<IGitInfoProvider>();
+            var repositoryMock = new Mock<IRepository>(MockBehavior.Loose);
+            var commitMock = new Mock<Commit>(MockBehavior.Loose);
+            var branchMock = new Mock<Branch>(MockBehavior.Strict);
+            var patchMock = new Mock<Patch>(MockBehavior.Strict);
+            var patchEntryChangesMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+            var patchEntryChangesGitIgnoreMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+
+            // Setup of mocks
+            commitMock
+                .SetupGet(x => x.Tree)
+                .Returns(new Mock<Tree>(MockBehavior.Loose).Object);
+
+            branchMock
+                .SetupGet(x => x.Tip)
+                .Returns(commitMock.Object);
+
+            branchMock.SetupGet(x => x.UpstreamBranchCanonicalName).Returns("origin/branch");
+            branchMock.SetupGet(x => x.CanonicalName).Returns("refs/heads/branch");
+            branchMock.SetupGet(x => x.FriendlyName).Returns("branch");
+
+            repositoryMock
+                .Setup(x => x.Branches.GetEnumerator())
+                .Returns(new List<Branch> { branchMock.Object }.GetEnumerator())
+                .Verifiable();
+
+            patchEntryChangesMock
+                .SetupGet(x => x.Path)
+                .Returns("file.cs");
+
+            patchEntryChangesGitIgnoreMock
+                .SetupGet(x => x.Path)
+                .Returns(FilePathUtils.NormalizePathSeparators($"{basePath}/Test.cs"));
+
+            patchMock
+                .Setup(x => x.GetEnumerator())
+                .Returns(((IEnumerable<PatchEntryChanges>)new List<PatchEntryChanges>
+                {
+                    patchEntryChangesMock.Object,
+                    patchEntryChangesGitIgnoreMock.Object
+                }).GetEnumerator());
+
+            repositoryMock
+                .Setup(x => x.Diff.Compare<Patch>(It.IsAny<Tree>(), DiffTargets.WorkingDirectory))
+                .Returns(patchMock.Object);
+
+            repositoryMock
+                .Setup(x => x.Lookup(It.IsAny<ObjectId>())).Returns(commitMock.Object);
+
+            gitInfoMock.Setup(x => x.DetermineCommit()).Returns(commitMock.Object);
+
+            gitInfoMock.SetupGet(x => x.Repository).Returns(repositoryMock.Object);
+            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns(FilePathUtils.NormalizePathSeparators("/c/Path/To/Repo"));
+            var target = new GitDiffProvider(options, gitInfoMock.Object);
+
+            // Act
+            var res = target.ScanDiff();
+
+            // Assert
+            res.ChangedTestFiles.Count().ShouldBe(0);
+            res.ChangedSourceFiles.Count().ShouldBe(1);
+        }
+
+        [Fact]
+        public void ScanDiffReturnsListofFiles_ExcludingTestFilesInDiffIgnoreFiles_Multi_Asterisk()
+        {
+            // Arrange
+            var diffIgnoreFiles = new string[1];
+            diffIgnoreFiles[0] = "**/Test.cs";
+
+            var basePath = FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests");
+            var options = new StrykerOptions(gitDiffTarget: "d670460b4b4aece5915caf5c68d12f560a9fe3e4", basePath: basePath, fileSystem: new MockFileSystem(), diffIgnoreFiles: diffIgnoreFiles);
+
+            var gitInfoMock = new Mock<IGitInfoProvider>();
+            var repositoryMock = new Mock<IRepository>(MockBehavior.Loose);
+            var commitMock = new Mock<Commit>(MockBehavior.Loose);
+            var branchMock = new Mock<Branch>(MockBehavior.Strict);
+            var patchMock = new Mock<Patch>(MockBehavior.Strict);
+            var patchEntryChangesMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+            var patchEntryChangesGitIgnoreMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+
+            // Setup of mocks
+            commitMock
+                .SetupGet(x => x.Tree)
+                .Returns(new Mock<Tree>(MockBehavior.Loose).Object);
+
+            branchMock
+                .SetupGet(x => x.Tip)
+                .Returns(commitMock.Object);
+
+            branchMock.SetupGet(x => x.UpstreamBranchCanonicalName).Returns("origin/branch");
+            branchMock.SetupGet(x => x.CanonicalName).Returns("refs/heads/branch");
+            branchMock.SetupGet(x => x.FriendlyName).Returns("branch");
+
+            repositoryMock
+                .Setup(x => x.Branches.GetEnumerator())
+                .Returns(new List<Branch> { branchMock.Object }.GetEnumerator())
+                .Verifiable();
+
+            patchEntryChangesMock
+                .SetupGet(x => x.Path)
+                .Returns("file.cs");
+
+            patchEntryChangesGitIgnoreMock
+                .SetupGet(x => x.Path)
+                .Returns(FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests/Test.cs"));
+
+            patchMock
+                .Setup(x => x.GetEnumerator())
+                .Returns(((IEnumerable<PatchEntryChanges>)new List<PatchEntryChanges>
+                {
+                    patchEntryChangesMock.Object,
+                    patchEntryChangesGitIgnoreMock.Object
+                }).GetEnumerator());
+
+            repositoryMock
+                .Setup(x => x.Diff.Compare<Patch>(It.IsAny<Tree>(), DiffTargets.WorkingDirectory))
+                .Returns(patchMock.Object);
+
+            repositoryMock
+                .Setup(x => x.Lookup(It.IsAny<ObjectId>())).Returns(commitMock.Object);
+
+            gitInfoMock.Setup(x => x.DetermineCommit()).Returns(commitMock.Object);
+
+            gitInfoMock.SetupGet(x => x.Repository).Returns(repositoryMock.Object);
+            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns(FilePathUtils.NormalizePathSeparators("/c/Path/To/Repo"));
+            var target = new GitDiffProvider(options, gitInfoMock.Object);
+
+            // Act
+            var res = target.ScanDiff();
+
+            // Assert
+            res.ChangedTestFiles.Count().ShouldBe(0);
+            res.ChangedSourceFiles.Count().ShouldBe(1);
+        }
+
+        [Fact]
+        public void ScanDiffReturnsListofFiles_ExcludingFilesInDiffIgnoreFiles_Multi_Asterisk()
+        {
+            // Arrange
+            var diffIgnoreFiles = new string[1];
+            diffIgnoreFiles[0] = "**/file.cs";
+
+            var basePath = FilePathUtils.NormalizePathSeparators("/c/Users/JohnDoe/Project/Tests");
+            var options = new StrykerOptions(gitDiffTarget: "d670460b4b4aece5915caf5c68d12f560a9fe3e4", basePath: basePath, fileSystem: new MockFileSystem(), diffIgnoreFiles: diffIgnoreFiles);
+
+            var gitInfoMock = new Mock<IGitInfoProvider>();
+            var repositoryMock = new Mock<IRepository>(MockBehavior.Loose);
+            var commitMock = new Mock<Commit>(MockBehavior.Loose);
+            var branchMock = new Mock<Branch>(MockBehavior.Strict);
+            var patchMock = new Mock<Patch>(MockBehavior.Strict);
+            var patchEntryChangesMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+            var patchEntryChangesGitIgnoreMock = new Mock<PatchEntryChanges>(MockBehavior.Strict);
+
+            // Setup of mocks
+            commitMock
+                .SetupGet(x => x.Tree)
+                .Returns(new Mock<Tree>(MockBehavior.Loose).Object);
+
+            branchMock
+                .SetupGet(x => x.Tip)
+                .Returns(commitMock.Object);
+
+            branchMock.SetupGet(x => x.CanonicalName).Returns("refs/heads/branch");
+
+            branchMock.SetupGet(x => x.FriendlyName).Returns("branch");
+
+            repositoryMock
+                .Setup(x => x.Branches.GetEnumerator())
+                .Returns(new List<Branch> { branchMock.Object }.GetEnumerator())
+                .Verifiable();
+
+            patchEntryChangesMock
+                .SetupGet(x => x.Path)
+                .Returns("/c/Users/JohnDoe/Project/file.cs");
+
+            patchMock
+                .Setup(x => x.GetEnumerator())
+                .Returns(((IEnumerable<PatchEntryChanges>)new List<PatchEntryChanges>
+                                                              {
+                                                                  patchEntryChangesMock.Object
+                                                              }).GetEnumerator());
+
+            repositoryMock
+                .Setup(x => x.Diff.Compare<Patch>(It.IsAny<Tree>(), DiffTargets.WorkingDirectory))
+                .Returns(patchMock.Object);
+
+            repositoryMock
+                .Setup(x => x.Lookup(It.IsAny<ObjectId>())).Returns(commitMock.Object);
+
+            gitInfoMock.Setup(x => x.DetermineCommit()).Returns(commitMock.Object);
+
+            gitInfoMock.SetupGet(x => x.Repository).Returns(repositoryMock.Object);
+            gitInfoMock.SetupGet(x => x.RepositoryPath).Returns("/c/Path/To/Repo");
+            var target = new GitDiffProvider(options, gitInfoMock.Object);
+
+            // Act
+            var res = target.ScanDiff();
+
+            // Assert
+            res.ChangedSourceFiles.Count().ShouldBe(0);
         }
     }
 }