+++ src/Stryker.Core/Stryker.Core/Options/StrykerOptions.cs
@@ -54,6 +54,8 @@ namespace Stryker.Core.Options
         public string ProjectVersion { get; }
         public MutationLevel MutationLevel { get; }
 
+        public IEnumerable<FilePattern> DiffIgnoreFiles { get; }
+
         public string AzureSAS { get; }
 
         public string AzureFileStorageUrl { get; set; }
@@ -101,7 +103,8 @@ namespace Stryker.Core.Options
             string azureSAS = null,
             string azureFileStorageUrl = null,
             IEnumerable<string> testProjects = null,
-            string mutationLevel = null)
+            string mutationLevel = null,
+            string[] diffIgnoreFiles = null)
         {
             _logger = logger;
             _fileSystem = fileSystem ?? new FileSystem();
@@ -132,7 +135,8 @@ namespace Stryker.Core.Options
             DashboardUrl = dashboardUrl;
             (DashboardApiKey, ProjectName) = ValidateDashboardReporter(dashboardApiKey, projectName);
             (ProjectVersion, FallbackVersion) = ValidateCompareToDashboard(projectVersion, fallbackVersion, gitDiffTarget);
-            ModuleName = moduleName;
+            DiffIgnoreFiles = ValidateDiffIgnoreFiles(diffIgnoreFiles);
+            ModuleName = !Reporters.Contains(Reporter.Dashboard) ? null : moduleName;
             BaselineProvider = ValidateBaselineProvider(baselineStorageLocation);
             (AzureSAS, AzureFileStorageUrl) = ValidateAzureFileStorage(azureSAS, azureFileStorageUrl);
             MutationLevel = ValidateMutationLevel(mutationLevel ?? MutationLevel.Standard.ToString());
@@ -254,6 +258,19 @@ namespace Stryker.Core.Options
             }
 
             return (projectVersion, fallbackVersion);
+        }
+
+        private IEnumerable<FilePattern> ValidateDiffIgnoreFiles(IEnumerable<string> diffIgnoreFiles)
+        {
+            var mappedDiffIgnoreFiles = new List<FilePattern>();
+            if (diffIgnoreFiles != null)
+            {
+                foreach (var pattern in diffIgnoreFiles)
+                {
+                    mappedDiffIgnoreFiles.Add(FilePattern.Parse(FilePathUtils.NormalizePathSeparators(pattern)));
+                }
+            }
+            return mappedDiffIgnoreFiles;
         }
 
         private static IEnumerable<Regex> ValidateIgnoredMethods(IEnumerable<string> methodPatterns)
@@ -421,7 +438,7 @@ namespace Stryker.Core.Options
 
             if (maxConcurrentTestRunners > safeProcessorCount)
             {
-                _logger?.LogWarning("Using {0} testrunners which is more than reccomended {1} for normal system operation. This can have an impact on performance.", maxConcurrentTestRunners, safeProcessorCount);
+                _logger?.LogWarning("Using {0} testrunners which is more than recommended {1} for normal system operation. This can have an impact on performance.", maxConcurrentTestRunners, safeProcessorCount);
             }
 
             if (maxConcurrentTestRunners == 1)
@@ -489,7 +506,7 @@ namespace Stryker.Core.Options
 
             foreach (var includePattern in filePatterns)
             {
-                filesToInclude.Add(FilePattern.Parse(includePattern));
+                filesToInclude.Add(FilePattern.Parse(FilePathUtils.NormalizePathSeparators(includePattern)));
             }
 
             if (filesToInclude.All(f => f.IsExclude))